projects/iptc-multi/serve/src/serve/served_model.py
"""Prediction model."""

# Standard Library
from typing import Type

# 3rd party libraries
from pydantic import BaseModel

# Internal libraries
from onclusiveml.serving.rest.serve import ServedModel

# Source
from src.serve.schemas import (
    BioResponseSchema,
    PredictRequestSchema,
    PredictResponseSchema,
)
from src.serve.topic_client import TopicClient
from src.settings import get_settings


settings = get_settings()


class ServedIPTCMultiModel(ServedModel):
    """Served IPTC Multi model."""

    predict_request_model: Type[BaseModel] = PredictRequestSchema
    predict_response_model: Type[BaseModel] = PredictResponseSchema
    bio_response_model: Type[BaseModel] = BioResponseSchema

    def __init__(self) -> None:
        super().__init__(name="iptc_multi")

    def load(self) -> None:
        """Load model."""
        self.model = TopicClient
        self.ready = True

    def predict(self, payload: PredictRequestSchema) -> PredictResponseSchema:
        """Topic Multi-model prediction.

        Args:
            payload (PredictRequestModel): prediction request payload.
        """
        # extract inputs data and inference specs from incoming payload
        inputs = payload.attributes
        configuration = payload.parameters

        iptc_list = self.model(input_data=inputs.content, **configuration.dict())
        return PredictResponseSchema.from_data(
            version=int(settings.api_version[1:]),
            namespace=settings.model_name,
            attributes={"iptc_topic": [iptc.dict() for iptc in iptc_list]},
        )

    def bio(self) -> BioResponseSchema:
        """Get bio information about the served IPTC Multi model.

        Returns:
            BioResponseSchema: Bio information about the model
        """
        return BioResponseSchema.from_data(
            version=int(settings.api_version[1:]),
            namespace=settings.model_name,
            attributes={"model_name": settings.model_name},
        )
